Add files via upload
Download Wireshark version 4.04, I have a Mac so I downloaded macOS Intel 64-bit.dmg. I did try macOS Arm 64-bit.dmg but it did not work, so be sure to download the correct version for your device. 

Packet Capture:
When you open Wireshark, there are a bunch of interfaces, you don't need any privileged administrator access to capture live network data so just choose the WiFi interface. 
Immediately, Wireshark will start capturing data and you can click the red square on the top left to stop the capture. To resume click the shark tail icon and a pop up will come up and ask if you'd like to save the packet capture, this is entirely up to you but, in this instance I did not; so I clicked 'continue without saving'. This is how you start and stop packet captures.